tests/mediacapture-depth/notadd/MediaStream-MediaElement-preload-none-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test that the HTMLMediaElement preload 'none' attribute value is ignored for MediaStream used as srcObject and MediaStream object URLs used as src.</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#mediastreams-in-media-elements">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Description</h2>
<p>
  This test checks that the HTMLMediaElement preload 'none' attribute value is ignored for MediaStream used as srcObject and MediaStream object URLs used as src.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream.
  </li>
</ol>

<div id="log"></div>

<video preload="none" id="video_src"></video>
<video preload="none" id="video_srcObject"></video>
<script>

function testPreloadNone(t, mediaElement, setSourceStreamFunc) {
  // The optional deferred load steps (for preload none) for MediaStream resources should be skipped.
  mediaElement.addEventListener("suspend", t.unreached_func("'suspend' should not be fired."));
  mediaElement.addEventListener("error", t.step_func(() => {
    assert_unreached("'error' should not be fired, code=" + mediaElement.error.code);
  }));

  mediaElement.addEventListener("loadeddata", t.step_func(() => {
    assert_equals(mediaElement.networkState, mediaElement.NETWORK_LOADING);
    t.done();
  }));

  setSourceStreamFunc();
  assert_equals(mediaElement.networkState, mediaElement.NETWORK_NO_SOURCE); // Resource selection is active.
}

async_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    let video_src = document.querySelector("#video_src");
    testPreloadNone(t, video_src, t.step_func(() => {
                        video_src.src = URL.createObjectURL(stream);
                        t.add_cleanup(() => { URL.revokeObjectURL(video_src.src); });
                    }));
  })
}, "Test that preload 'none' is ignored for MediaStream object URL used as src");

async_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    let video_srcObject = document.querySelector("#video_srcObject");
    testPreloadNone(t, video_srcObject, t.step_func(() => { video_srcObject.srcObject = stream; }));
  })
}, "Test that preload 'none' is ignored for MediaStream used as srcObject");

</script>
